refactor(cli): extract handoff capability owner
